Switched to git repo for crate
Git version of the crate seems to more reliably provide the steamworks SDK, which will cause strange errors at startup if not prooperly configured. Targeting the same specific tag that the version of crates.io so this shouldn't cause unecessary problems. Manual upgrades will need to be done over time. At this point I might just be willing to maintain this crate since it's so useful for publishing games on steam!